studio/requirements.txt
langgraph~=0.6.4
pandas~=2.3.1
typing_extensions~=4.14.1
langchain-core~=0.3.74
pydantic~=2.11.7
pandasql~=0.7.3
langchain-openai~=0.3.29
openai~=1.99.6
dotenv~=0.9.9
python-dotenv~=1.1.1
matplotlib~=3.10.5
reportlab~=4.4.3
altair~=5.5.0
seaborn~=0.13.2
numpy~=2.3.2
plotly~=6.3.0
networkx~=2.8.3
PyYAML~=6.0.2